Refactor `ToggleGroupControl` folder structure
