+++ packages/components/src/toggle-group-control/index.ts
@@ -1,2 +1,2 @@
-export { default as ToggleGroupControl } from './component';
-export { default as ToggleGroupControlOption } from './toggle-group-control-option';
+export { ToggleGroupControl } from './toggle-group-control';
+export { ToggleGroupControlOption } from './toggle-group-control-option';
