+++ packages/components/src/toggle-group-control/toggle-group-control-option/README.md
@@ -1,0 +1,40 @@
+# `ToggleGroupControlOption`
+
+<div class="callout callout-alert">
+This feature is still experimental. “Experimental” means this is an early implementation subject to drastic and breaking changes.
+</div>
+
+`ToggleGroupControlOption` is a form component and is meant to be used as a child of [`ToggleGroupControl`]((/packages/components/src/toggle-group-control/toggle-group-control/README.md)).
+
+
+## Usage
+
+```js
+import {
+	__experimentalToggleGroupControl as ToggleGroupControl,
+	__experimentalToggleGroupControlOption as ToggleGroupControlOption,
+} from '@wordpress/components';
+
+function Example() {
+	return (
+		<ToggleGroupControl label="my label" value="vertical" isBlock>
+			<ToggleGroupControlOption value="horizontal" label="Horizontal" />
+			<ToggleGroupControlOption value="vertical" label="Vertical" />
+		</ToggleGroupControl>
+	);
+}
+```
+
+## Props
+
+### `label`: `string`
+
+Label for the option. If needed, the `aria-label` prop can be used in addition to specify a different label for assistive technologies.
+
+- Required: Yes
+
+### `value`: `string | number`
+
+The value of the `ToggleGroupControlOption`.
+
+- Required: Yes
