+++ packages/components/src/toggle-group-control/toggle-group-control-option/index.ts
@@ -1,0 +1,1 @@
+export { default as ToggleGroupControlOption } from './component';
