+++ packages/components/src/toggle-group-control/toggle-group-control/index.ts
@@ -1,0 +1,1 @@
+export { default as ToggleGroupControl } from './component';
